fixup move get_terminal_columns() to terminal.c
