+++ src/terminal.h
@@ -1,5 +1,5 @@
-#ifndef HEADER_CURL_TOOL_CB_PRG_H
-#define HEADER_CURL_TOOL_CB_PRG_H
+#ifndef HEADER_CURL_TERMINAL_H
+#define HEADER_CURL_TERMINAL_H
 /***************************************************************************
  *                                  _   _ ____  _
  *  Project                     ___| | | |  _ \| |
@@ -25,32 +25,6 @@
  ***************************************************************************/
 #include "tool_setup.h"
 
-#define CURL_PROGRESS_STATS 0 /* default progress display */
-#define CURL_PROGRESS_BAR   1
+unsigned int get_terminal_columns(void);
 
-struct ProgressData {
-  int         calls;
-  curl_off_t  prev;
-  struct timeval prevtime;
-  int         width;
-  FILE       *out;  /* where to write everything to */
-  curl_off_t  initial_size;
-  unsigned int tick;
-  int bar;
-  int barmove;
-};
-
-struct OperationConfig;
-
-void progressbarinit(struct ProgressData *bar,
-                     struct OperationConfig *config);
-
-/*
-** callback for CURLOPT_PROGRESSFUNCTION
-*/
-
-int tool_progress_cb(void *clientp,
-                     curl_off_t dltotal, curl_off_t dlnow,
-                     curl_off_t ultotal, curl_off_t ulnow);
-
-#endif /* HEADER_CURL_TOOL_CB_PRG_H */
+#endif /* HEADER_CURL_TERMINAL_H */
