+++ src/tool_cb_prg.c
@@ -23,10 +23,6 @@
  ***************************************************************************/
 #include "tool_setup.h"
 
-#ifdef HAVE_SYS_IOCTL_H
-#include <sys/ioctl.h>
-#endif
-
 #define ENABLE_CURLX_PRINTF
 /* use our own printf() functions */
 #include "curlx.h"
@@ -35,16 +31,11 @@
 #include "tool_cb_prg.h"
 #include "tool_util.h"
 #include "tool_operate.h"
+#include "terminal.h"
 
 #include "memdebug.h" /* keep this as LAST include */
 
 #define MAX_BARLENGTH 256
-
-#ifdef HAVE_TERMIOS_H
-#  include <termios.h>
-#elif defined(HAVE_TERMIO_H)
-#  include <termio.h>
-#endif
 
 /* 200 values generated by this perl code:
 
@@ -227,59 +218,6 @@ int tool_progress_cb(void *clientp,
   }
 
   return 0;
-}
-
-/*
- * get_terminal_columns() returns the number of columns in the current
- * terminal. It will return 79 on failure. Also, the number can be very big.
- */
-
-unsigned int get_terminal_columns(void)
-{
-  unsigned int width = 0;
-  char *colp = curl_getenv("COLUMNS");
-  if(colp) {
-    char *endptr;
-    long num = strtol(colp, &endptr, 10);
-    if((endptr != colp) && (endptr == colp + strlen(colp)) && (num > 20) &&
-       (num < 10000))
-      width = (unsigned int)num;
-    curl_free(colp);
-  }
-
-  if(!width) {
-    int cols = 0;
-
-#ifdef TIOCGSIZE
-    struct ttysize ts;
-    if(!ioctl(STDIN_FILENO, TIOCGSIZE, &ts))
-      cols = ts.ts_cols;
-#elif defined(TIOCGWINSZ)
-    struct winsize ts;
-    if(!ioctl(STDIN_FILENO, TIOCGWINSZ, &ts))
-      cols = (int)ts.ws_col;
-#elif defined(_WIN32)
-    {
-      HANDLE  stderr_hnd = GetStdHandle(STD_ERROR_HANDLE);
-      CONSOLE_SCREEN_BUFFER_INFO console_info;
-
-      if((stderr_hnd != INVALID_HANDLE_VALUE) &&
-         GetConsoleScreenBufferInfo(stderr_hnd, &console_info)) {
-        /*
-         * Do not use +1 to get the true screen-width since writing a
-         * character at the right edge will cause a line wrap.
-         */
-        cols = (int)
-          (console_info.srWindow.Right - console_info.srWindow.Left);
-      }
-    }
-#endif /* TIOCGSIZE */
-    if(cols >= 0 && cols < 10000)
-      width = (unsigned int)cols;
-  }
-  if(!width)
-    width = 79;
-  return width; /* 79 for unknown, might also be very small or very big */
 }
 
 void progressbarinit(struct ProgressData *bar,
